Working branch (#43)
* Started kidnapped princess

* princess layla bugfix and more

* fix some crashes

* Holy!

* Fixes

* fix equipment dealer

* Minor fixes

* minor refactorings

* Refactoring module registration

* Typos

* fixes, refactorings

* Update modularization

* Update dungeon generator

* Changes Cave

* added FishingVillage

* prepare merge

* FishinVillge practically done

* typo

* user file entfernt

* Addedm Questlog

* even more fishing

* delete CMakeLists.txt.user

* Kleinigkeiten

* Bugfix

* kleinigkeiten

* refactoring mapsymbol

* Cave now integrated in the game progress. expanded tasks

* fixes

* LootItems refactoring

* refactoring itemfactory ctd.

* DungeonMap

* Refactorings, fixes

* shaggys sandwich

* sandwich

* minor restructure, enumiterator

* refactorings

* Kleinigkeiten

* added sandwich shop

* Sandwichshop ctd.

* even more sandwichstore

* Bugfix

* Fixes

* sandwichshop fixes

* fix

* another fix

* Added Hint System

* changed handling of golf.

* Added Layla Pt. 2

* Bugfixes

* small fixes

* Started Rebellion

* Text Updates

* Game Progression update

* fix in sandwichShop

* fixes

* balancing

* kleikram

* added gitignore

* this is really bad!

* Just to make sure

* kleinigkeiten, balancing

* fuckup beim merge

* activated next chapter

* Refactorings

* kleinkram

* kleinscheiß

* Major Restructuring of source

* Major Restructuring of code

* repair restructuring

* merh umtüddeln

* added Questsystem

* umbau sandwichshop

* kleines refactoring

* startet black ivory tower, shrine as module

* redesign shrine as module, started black ivory tower

* Tower of the lunatics

* merge, minor changes

* fix deployment

* fix deplyment

* fix deploy,e

* fix warnings

* ...

* fix build

* get rid of warnings

* fix build

* fix build

* warnings

+++ core/cgameprogression.cpp
@@ -236,7 +236,7 @@ void CGameProgression::increaseTurns()
 
 unsigned int CGameProgression::progress() const
 {
-    return std::ceil((_finishedModules.size() * 100) / _registeredModules.size());
+    return static_cast<unsigned int>(std::ceil((_finishedModules.size() * 100) / _registeredModules.size()));
 }
 
 unsigned long CGameProgression::bodyCount() const
